Fixes a bug with importing from the java standard library.
-- The java standard library has packages `java.io` and `java.net`.
Unfortunately, `io` and `net` are also valid package starts. *Furthermore* we
can't just find the package from the start of the directory, because mvn-based
projects have a directory structure like `src/main/java`, which java shouldn't
be part of the package.

-- So, we implemented a hack specific to my machine. If the tag is under
"java-standard-libary" or "open-jdk" then it removes "net" and "io" from the
map of root packages before we start searching for the leading package.

-- This could probably be generalized to make it slightly less hacky for other
machines, but it's unfortunate this won't "just work."

-- Perhaps at some point we could implement some discover code? Basically,
look for "java/io/IOException" and use the directory just *before* that to
as the "this is the sdk" directory? Meh, that sounds too magical. Better just
to let people do it.