Omni completion is now fast.
-- First, we now build only the bare minimum dictionary for
the completions (the function signature). This saves us some
significant time because we don't have to load all the files into
memory to get their package (note to self: optimize this).

-- Second, we tweak how we get the signature so that now we're
calling out to grep, rather than loading the whole file into
memory. Instead, we only load the twenty lines of each match
in the file (which will likely only be one or two) to ensure
we get the full signature in all but the most pathological of
cases.

-- And now, I don't get any delay where before I got a small one!